Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Du",email = "duducostalobo10@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.118.0,<0.119.0)",
    "pydantic[email] (>=2.11.9,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.44,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "tzdata (>=2025.2,<2026.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "alembic (>=1.17.0,<2.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "langchain (>=1.0.2,<2.0.0)",
    "langchain-core (>=1.0.1,<2.0.0)",
    "langchain-community (>=0.4,<0.5)",
    "langchain-groq (>=1.0.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "langgraph (>=1.0.1,<2.0.0)",
    "fastapi-profiler (>=1.4.1,<2.0.0)",
    "psycopg[binary] (>=3.2.12,<4.0.0)",
    "langchain-openai (>=1.0.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.13.2,<0.14.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "freezegun (>=1.5.5,<2.0.0)"
]


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev backend/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=Backend -vv'
post_test = 'coverage html'